Added feedback viewmodel and state handling

+++ app/src/main/java/uk/ac/tees/mad/foodreview/domain/model/Review.kt
@@ -5,7 +5,7 @@ import android.media.Rating
 data class Review(
     val id: String ="",
     val foodName : String = "",
-    val rating: Double = 0.0,
+    val rating: Int = 0,
     val reviewText : String = "",
     val userId : String = "",
     val createdAt : Long = System.currentTimeMillis()
